third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/hard_swish.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] hardSwish float16 0D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test hardSwish float16 actual 0 should be close enough to expected 0.45703125 by ULP distance: expected a number less than or equal to 4 but got 14160
[FAIL] hardSwish float16 1D constant tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test hardSwish float16 actual 0 should be close enough to expected 0.45703125 by ULP distance: expected a number less than or equal to 4 but got 14160
[FAIL] hardSwish float16 1D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test hardSwish float16 actual 0 should be close enough to expected 0.45703125 by ULP distance: expected a number less than or equal to 4 but got 14160
[FAIL] hardSwish float16 2D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test hardSwish float16 actual 0 should be close enough to expected 0.45703125 by ULP distance: expected a number less than or equal to 4 but got 14160
[FAIL] hardSwish float16 3D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test hardSwish float16 actual 0 should be close enough to expected 0.45703125 by ULP distance: expected a number less than or equal to 4 but got 14160
[FAIL] hardSwish float16 4D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test hardSwish float16 actual 0 should be close enough to expected 0.45703125 by ULP distance: expected a number less than or equal to 4 but got 14160
Harness: the test ran to completion.
